+++ ZwiftViewer/README.md
@@ -1,17 +1,34 @@
-This is a Kotlin Multiplatform project targeting Android, iOS.
+# ZwiftViewer (Kotlin Multiplatform)
 
-* `/composeApp` is for code that will be shared across your Compose Multiplatform applications.
-  It contains several subfolders:
-    - `commonMain` is for code that’s common for all targets.
-    - Other folders are for Kotlin code that will be compiled for only the platform indicated in the folder name.
-      For example, if you want to use Apple’s CoreCrypto for the iOS part of your Kotlin app,
-      `iosMain` would be the right folder for such calls.
+ZwiftViewer is a Kotlin Multiplatform (KMP) mobile app built to showcase shared business logic, UI patterns, and network integration across Android and iOS. It demonstrates the foundation of a cross-platform companion app experience for a Zwift-like cycling platform.
 
-* `/iosApp` contains iOS applications. Even if you’re sharing your UI with Compose Multiplatform,
-  you need this entry point for your iOS app. This is also where you should add SwiftUI code for your project.
+## Features
 
-* `/shared` is for the code that will be shared between all targets in the project.
-  The most important subfolder is `commonMain`. If preferred, you can add code to the platform-specific folders here
-  too.
+- Cross-platform login and session management
+- Real-time ride data retrieval from ZwiftPower
+- Shared rendering of ride summaries and stats
+- Modern, native UI using Jetpack Compose (Android) and SwiftUI (iOS)
 
-Learn more about [Kotlin Multiplatform](https://www.jetbrains.com/help/kotlin-multiplatform-dev/get-started.html)…
+## Tech Stack
+
+- Kotlin Multiplatform Mobile (KMM)
+- Ktor (HTTP client)
+- Kotlinx Serialization
+- SwiftUI
+- Jetpack Compose
+- Gradle (build)
+- XCFramework (iOS integration)
+
+## Project Structure
+
+- `shared/` — Shared Kotlin module (models, networking, business logic)
+- `composeApp/` — Android app (Jetpack Compose)
+- `iosApp/` — iOS app (SwiftUI + integrated shared Kotlin framework)
+
+## Resources
+
+- [Kotlin Multiplatform Docs](https://www.jetbrains.com/help/kotlin-multiplatform-dev/get-started.html)
+
+---
+
+This project was developed as a demonstration of mobile engineering capability for a cross-platform cycling app.
