+++ ZwiftViewer/composeApp/src/androidMain/kotlin/com/danielchew/zwiftviewer/ui/login/LoginScreen.kt
@@ -1,57 +1,66 @@
 package com.danielchew.zwiftviewer.ui.login
 
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Button
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
+import android.webkit.WebView
+import android.webkit.WebViewClient
+import android.webkit.CookieManager
+import android.content.Context
+import android.view.ViewGroup
+import androidx.compose.foundation.layout.Box
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.navigation.NavController
-import com.danielchew.zwiftviewer.ui.login.LoginViewModel
+import androidx.compose.ui.viewinterop.AndroidView
 
 @Composable
-fun LoginScreen(navController: NavController) {
-    val viewModel: LoginViewModel = viewModel()
-    val uiState by viewModel.uiState.collectAsState()
+fun ZwiftPowerLoginScreen(
+    onCookiesExtracted: (Map<String, String>) -> Unit,
+    onPageUrlChanged: (String) -> Unit,
+    onLoginSuccess: () -> Unit,
+    modifier: Modifier,
+    urlToLoad: String
+) {
+    Box(modifier = modifier) {
+        AndroidView(
+            factory = { context: Context ->
+                WebView(context).apply {
+                    layoutParams = ViewGroup.LayoutParams(
+                        ViewGroup.LayoutParams.MATCH_PARENT,
+                        ViewGroup.LayoutParams.MATCH_PARENT
+                    )
+                    settings.javaScriptEnabled = true
+                    webViewClient = object : WebViewClient() {
+                        override fun onPageFinished(view: WebView?, url: String?) {
+                            if (url == null) {
+                                println("ZwiftDebug: onPageFinished called with null URL.")
+                                return
+                            }
 
-    Column(modifier = Modifier.fillMaxSize().padding(16.dp)) {
-        Text(text = "Login to Zwift", style = MaterialTheme.typography.headlineSmall)
-        Spacer(modifier = Modifier.height(16.dp))
+                            println("ZwiftDebug: Page finished loading: $url")
 
-        Text(text = "Username")
-        TextField(
-            value = uiState.username,
-            onValueChange = { viewModel.onUsernameChanged(it) },
-            modifier = Modifier.fillMaxWidth()
+                            val cookieManager = CookieManager.getInstance()
+                            val rawCookies = cookieManager.getCookie(url)
+                            if (rawCookies.isNullOrBlank()) {
+                                println("ZwiftDebug: No cookies found for $url")
+                                return
+                            }
+
+                            val parsedCookies = rawCookies
+                                .split(";")
+                                .mapNotNull { it.trim().split("=").takeIf { it.size == 2 } }
+                                .associate { it[0] to it[1] }
+
+                            println("ZwiftDebug: Parsed cookie keys: ${parsedCookies.keys}")
+                            onCookiesExtracted(parsedCookies)
+
+                            if (url.contains("profile.php")) {
+                                println("ZwiftDebug: Detected profile.php – assuming login success.")
+                                onLoginSuccess()
+                            }
+                        }
+                    }
+                    loadUrl(urlToLoad)
+                }
+            },
+            modifier = modifier
         )
-
-        Spacer(modifier = Modifier.height(8.dp))
-
-        Text(text = "Password")
-        TextField(
-            value = uiState.password,
-            onValueChange = { viewModel.onPasswordChanged(it) },
-            modifier = Modifier.fillMaxWidth()
-        )
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        Button(onClick = {
-            if (viewModel.onLoginClicked()) {
-                navController.navigate("rides")
-            }
-        }) {
-            Text("Login")
-        }
     }
 }
