--- ZwiftViewer/composeApp/src/androidMain/kotlin/com/danielchew/zwiftviewer/ui/login/LoginUiState.kt
@@ -1,6 +1,0 @@
-package com.danielchew.zwiftviewer.ui.login
-
-data class LoginUiState(
-    val username: String = "",
-    val password: String = ""
-)
