--- ZwiftViewer/shared/src/androidMain/kotlin/com/danielchew/zwiftviewer/network/HttpClientEngineProvider.kt
@@ -1,6 +1,0 @@
-package com.danielchew.zwiftviewer.network
-
-import io.ktor.client.engine.HttpClientEngineFactory
-import io.ktor.client.engine.cio.CIO
-
-actual fun getPlatformEngine(): HttpClientEngineFactory<*> = CIO
