--- ZwiftViewer/shared/src/androidMain/kotlin/com/danielchew/zwiftviewer/utils/CookieUtils.kt
@@ -1,22 +1,0 @@
-package com.danielchew.zwiftviewer.utils
-
-import android.util.Log
-import android.webkit.CookieManager
-
-actual fun extractCookiesForDomain(domain: String): Map<String, String> {
-    val cookieManager = CookieManager.getInstance()
-    val rawCookies = cookieManager.getCookie(domain) ?: return emptyMap()
-
-    val parsedCookies = rawCookies.split(";").mapNotNull {
-        val trimmed = it.trim()
-        val parts = trimmed.split("=", limit = 2)
-        if (parts.size == 2 && parts[0].isNotBlank()) {
-            parts[0] to parts[1]
-        } else {
-            null // skip malformed cookies
-        }
-    }.toMap()
-
-    Log.d("ZwiftDebug", "Parsed cookies for $domain: ${parsedCookies.keys}")
-    return parsedCookies
-}
