--- ZwiftViewer/shared/src/commonMain/kotlin/com/danielchew/zwiftviewer/Greeting.kt
@@ -1,9 +1,0 @@
-package com.danielchew.zwiftviewer
-
-class Greeting {
-    private val platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
