--- ZwiftViewer/shared/src/commonMain/kotlin/com/danielchew/zwiftviewer/utils/CookieUtils.kt
@@ -1,3 +1,0 @@
-package com.danielchew.zwiftviewer.utils
-
-expect fun extractCookiesForDomain(domain: String): Map<String, String>
