--- ZwiftViewer/shared/src/iosMain/kotlin/com/danielchew/zwiftviewer/network/HttpClientEngineProvider.kt
@@ -1,9 +1,0 @@
-package com.danielchew.zwiftviewer.network
-
-import io.ktor.client.engine.HttpClientEngineFactory
-import io.ktor.client.engine.darwin.Darwin
-
-actual fun getPlatformEngine(): HttpClientEngineFactory<*> {
-    println("ZwiftDebug: iOS engine = Darwin")
-    return Darwin
-}
