--- ZwiftViewer/shared/src/iosMain/kotlin/com/danielchew/zwiftviewer/utils/CookieUtils.kt
@@ -1,6 +1,0 @@
-package com.danielchew.zwiftviewer.utils
-
-actual fun extractCookiesForDomain(domain: String): Map<String, String> {
-    // iOS doesn't support this yet
-    return emptyMap()
-}
